feat(node-sdk): move logs api types to own package
