+++ node-sdk/lib/logstf-api-types/package.json
@@ -1,0 +1,12 @@
+{
+  "name": "logstf-api-types",
+  "version": "0.0.0",
+  "license": "MIT",
+  "main": "src/index.ts",
+  "types": "src/index.ts",
+  "devDependencies": {
+    "@types/node": "^18.11.17",
+    "tsup": "^6.5.0",
+    "typescript": "^4.9.4"
+  }
+}
