+++ node-sdk/lib/logstf-api-types/src/types/endpoints/logUploadRequest.ts
@@ -1,4 +1,4 @@
-export type LogUploadConfig = {
+export type LogUploadRequest = {
   title: string;
   map: string | null;
   uploader: string | null;
