+++ node-sdk/lib/logstf-api-types/src/types/endpoints/logUploadResponse.ts
