+++ node-sdk/lib/logstf-api-types/src/types/enums/teamType.ts
@@ -1,0 +1,4 @@
+export enum TeamType {
+  Red,
+  Blue,
+}
