+++ node-sdk/lib/logstf-api-types/src/types/medicStats.ts
@@ -1,0 +1,10 @@
+export type MedicStats = {
+  advantages_lost: number;
+  biggest_advantage_lost: number;
+  deaths_with_95_99_uber: number;
+  deaths_within_20s_after_uber: number;
+  avg_time_before_healing: number;
+  avg_time_to_build: number;
+  avg_time_before_using: number;
+  avg_uber_length: number;
+};
