+++ node-sdk/lib/logstf-api-types/src/types/roundEvent.ts
@@ -1,0 +1,8 @@
+import type { TeamType } from "./enums/teamType";
+
+export type RoundEvent = {
+  type: string;
+  time: number;
+  team: TeamType;
+  point: number;
+};
