+++ node-sdk/lib/logstf-api-types/src/types/teamInformation.ts
@@ -1,0 +1,10 @@
+export type TeamInformation = {
+  score: number;
+  kills: number;
+  deaths: number;
+  dmg: number;
+  charges: number;
+  drops: number;
+  firstcaps: number;
+  caps: number;
+};
