+++ node-sdk/lib/logstf-api-types/src/types/teamRoundInfo.ts
@@ -1,0 +1,6 @@
+export type TeamRoundInfo = {
+  score: number;
+  kills: number;
+  dmg: number;
+  ubers: number;
+};
