+++ node-sdk/lib/logstf-api-types/src/types/tf2ClassMap.ts
@@ -1,0 +1,11 @@
+export type Tf2ClassMap = {
+  scout?: number;
+  solider?: number;
+  pyro?: number;
+  demoman?: number;
+  heavyweapons?: number;
+  engineer?: number;
+  medic?: number;
+  sniper?: number;
+  spy?: number;
+};
