+++ node-sdk/lib/logstf-api-types/tsconfig.json
@@ -1,0 +1,15 @@
+{
+  "extends": "@sapphire/ts-config",
+  "compilerOptions": {
+    "rootDir": "src",
+    "module": "ESNext",
+    "target": "ESNext",
+    "outDir": "dist",
+    "emitDecoratorMetadata": false,
+    "tsBuildInfoFile": "dist/.tsbuildinfo",
+    "skipLibCheck": true,
+    "baseUrl": "src",
+    "lib": ["ESNext"],
+    "removeComments": true
+  }
+}
